refactor: Console.js 삭제
입출력을 InputView와 OutputView에서 이미 관리하는데
굳이 콘솔 입출력을 또 다른 곳에서 관리할 필요가 없다는 생각이 들어
Console을 삭제하고, 입력 처리 함수를 InputView 내부로 이동
